:wrench: Adjust to be align with Edrean's Directions
TL;DR;
Refactor to comply with PR#1 .
Change the Name of the Node, Refactor to accept an action.

Detailed Version:
Since the vision is have only 1 node per action, I refactor the node to accept the action so in the future this node can be complemented(e.g. Get Specific alert)